Refactor result ordering to use memoized selectors
